Experience table as a Tensor
In the original implementation the experience table was stored as a Lua table. This meant that we hit the 1GB memory map limit of LuaJIT even for moderate table sizes and replayed memories count (temporal window). 

I store the experience table as a Tensor, alleviating the memory limit and improving performance a bit.